feat: add schemars feature for json schemas
