+++ Cargo.toml
@@ -11,6 +11,7 @@ repository = "https://github.com/famedly/famedly-rust-utils"
 
 [dependencies]
 reqwest = { version = "0.12.12", optional = true }
+schemars = { version = "0.8.22", optional = true }
 serde = { version = "1.0.210", features = ["derive"] }
 thiserror = "1.0.64"
 time = { version = "0.3.36", optional = true }
@@ -23,6 +24,7 @@ serde_json = "1.0.128"
 [features]
 reqwest = ["dep:reqwest"]
 time = ["dep:time"]
+schemars = ["dep:schemars", "schemars/url"]
 
 [lints.rust]
 dead_code = "warn"
